+++ jsonutil/example_test.go
@@ -35,3 +35,49 @@ func ExampleDecodeFile() {
 		fmt.Printf("Error: %v\n", err)
 	}
 }
+
+func ExampleWrite() {
+	var data = make(map[string]int)
+
+	data["john"] = 100
+	data["bob"] = 300
+
+	err := Write("/path/to/file.json", data, 0600)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+	}
+}
+
+func ExampleRead() {
+	var data = make(map[string]int)
+
+	err := Read("/path/to/file.json", data)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+	}
+}
+
+func ExampleWriteGz() {
+	var data = make(map[string]int)
+
+	data["john"] = 100
+	data["bob"] = 300
+
+	err := WriteGz("/path/to/file.gz", data, 0600)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+	}
+}
+
+func ExampleReadGz() {
+	var data = make(map[string]int)
+
+	err := ReadGz("/path/to/file.gz", data)
+
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+	}
+}